+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2
 ### Added
 - Added zoom sliders for Outputs, Checkpoints, and Model Browser pages
 - Added toggle to hide "Early Access" models in the CivitAI Model Browser
+- Added trigger words to the Inference Extra Networks (Lora/Lyco) selector for quick copy & paste
 
 ## v2.12.0-pre.3
 ### Added